backend/pyproject.toml
[tool.poetry]
name = "minerva"
version = "0.1.0"
description = ""
authors = ["Piotr Gerke <gerke.contact@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
certifi = "^2024.2.2"
openai = "^1.23.3"
python-multipart = "^0.0.9"
tenacity = "^8.2.3"
path = "^16.14.0"
tiktoken = "^0.6.0"
arrow = "^1.3.0"
pydantic = "^2.7.1"
asyncpg = "^0.29.0"
greenlet = "^3.1.1"
email-validator = "^2.2.0"
aiohttp = "^3.10.10"
pymongo = "4.9.0"
motor = "3.6.0"
passlib = "^1.7.4"
pyjwt = "^2.9.0"
bcrypt = "^4.2.0"
playwright = "^1.48.0"
bs4 = "^0.0.2"
beautifulsoup4 = "^4.12.3"
ffmpeg = "^1.4"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
tabulate = "^0.9.0"
jsonschema = "^4.23.0"
pinecone = "^5.4.1"
pypdf = "^5.1.0"
python-docx = "^1.1.2"
pdfreader = "^0.1.15"
aiofiles = "^24.1.0"
apscheduler = "^3.11.0"
resend = "^2.4.0"
stripe = "^11.3.0"
google-auth = "^2.37.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
holidays = "^0.63"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.1"
urllib3 = "<2.0"
pdf2image = "^1.17.0"
pytesseract = "^0.3.13"
pdfplumber = "^0.11.5"
playwright-stealth = "^1.0.6"
setuptools = "*" 
celery = "^5.4.0"
redis = "^5.2.1"
vercel-blob = "^0.3.0"
anthropic = "^0.49.0"
nltk = "^3.9.1"
json-repair = "^0.40.0"
httpx = "^0.28.1"
jinja2 = "^3.1.6"
elasticsearch = "^8.17.2"
odfpy = "^1.4.1"
py7zr = "^0.22.0"
rarfile = "^4.2"
boto3 = "^1.34.0"

[tool.poetry.scripts]
start = "minerva.api.main:start"
dev = "minerva.config.run_dev:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
